iot_api_requirements.txt
# IoT Device REST API Dependencies
# Install with: pip install -r iot_api_requirements.txt

# Django REST Framework - API framework
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# API Documentation
drf-yasg==1.21.7  # Swagger/OpenAPI documentation

# CORS Headers (if devices connect from different origins)
django-cors-headers==4.3.0

# API Throttling and Rate Limiting
django-ratelimit==4.1.0

# JSON Web Tokens
cryptography==41.0.7

# API Testing
httpie==3.2.2  # Command-line HTTP client for testing
